Modernize example projects for ASP.NET Core 6.0
