+++ examples/AspNetCore/OData/ODataAdvancedExample/Properties/launchSettings.json
@@ -3,23 +3,25 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:1237/",
-      "sslPort": 0
+      "applicationUrl": "http://localhost:21237/",
+      "sslPort": 44328
     }
   },
   "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
+    "ODataAdvancedExample": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
       "launchBrowser": true,
-      "launchUrl": "api",
+      "launchUrl": "api/People?api-version=1.0",
+      "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "ODataAdvancedExample": {
-      "commandName": "Project",
+    "IIS Express": {
+      "commandName": "IISExpress",
       "launchBrowser": true,
-      "launchUrl": "http://localhost:5000/api",
+      "launchUrl": "api/People?api-version=1.0",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
