+++ examples/AspNetCore/OData/ODataBasicExample/Properties/launchSettings.json
@@ -4,22 +4,24 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:1238/",
-      "sslPort": 0
+      "sslPort": 44328
     }
   },
   "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
+    "ODataBasicExample": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
       "launchBrowser": true,
       "launchUrl": "api/People?api-version=1.0",
+      "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "ODataBasicExample": {
-      "commandName": "Project",
+    "IIS Express": {
+      "commandName": "IISExpress",
       "launchBrowser": true,
-      "launchUrl": "http://localhost:5000/api/$metadata",
+      "launchUrl": "api/People?api-version=1.0",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
