+++ examples/AspNetCore/WebApi/ByNamespaceExample/Properties/launchSettings.json
@@ -4,22 +4,24 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:26458/",
-      "sslPort": 0
+      "sslPort": 44328
     }
   },
   "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
+    "ByNamespaceExample": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
       "launchBrowser": true,
       "launchUrl": "v1/orders/42",
+      "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "ByNamespaceExample": {
-      "commandName": "Project",
+    "IIS Express": {
+      "commandName": "IISExpress",
       "launchBrowser": true,
-      "launchUrl": "http://localhost:5000/v1/orders/42",
+      "launchUrl": "v1/orders/42",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
