+++ examples/AspNetCore/WebApi/ConventionsExample/Properties/launchSettings.json
@@ -3,23 +3,25 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:2645/",
-      "sslPort": 0
+      "applicationUrl": "http://localhost:29645/",
+      "sslPort": 44328
     }
   },
   "profiles": {
-    "IIS Express": {
-      "commandName": "IISExpress",
+    "ConventionsExample": {
+      "commandName": "Project",
+      "dotnetRunMessages": true,
       "launchBrowser": true,
       "launchUrl": "api/values?api-version=1.0",
+      "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "ConventionsExample": {
-      "commandName": "Project",
+    "IIS Express": {
+      "commandName": "IISExpress",
       "launchBrowser": true,
-      "launchUrl": "http://localhost:5000/api/values?api-version=1.0",
+      "launchUrl": "api/values?api-version=1.0",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
