+++ examples/AspNetCore/WebApi/MinimalApiExample/Properties/launchSettings.json
@@ -5,7 +5,7 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:31998",
-      "sslPort": 0
+      "sslPort": 44328
     }
   },
   "profiles": {
@@ -14,7 +14,7 @@
       "dotnetRunMessages": true,
       "launchBrowser": true,
       "launchUrl": "weatherforecast?api-version=1.0",
-      "applicationUrl": "http://localhost:5144",
+      "applicationUrl": "https://localhost:5145;http://localhost:5144",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
