+++ examples/AspNetCore/WebApi/OpenApiExample/OpenApiExample.csproj
@@ -2,11 +2,10 @@
 
  <PropertyGroup>
   <TargetFramework>net6.0</TargetFramework>
-  <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(MSBuildThisFileName).xml</DocumentationFile>
+  <GenerateDocumentationFile>true</GenerateDocumentationFile>
  </PropertyGroup>
 
  <ItemGroup>
-  <PackageReference Include="Microsoft.Extensions.PlatformAbstractions" Version="1.1.0" />
   <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.0-*" />
  </ItemGroup>
