--- examples/AspNetCore/WebApi/OpenApiExample/Startup.cs
@@ -1,109 +1,0 @@
-﻿[assembly: Microsoft.AspNetCore.Mvc.ApiController]
-
-namespace ApiVersioning.Examples;
-
-using Asp.Versioning;
-using Asp.Versioning.ApiExplorer;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
-using Microsoft.Extensions.PlatformAbstractions;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using System.IO;
-using System.Reflection;
-
-/// <summary>
-/// Represents the startup process for the application.
-/// </summary>
-public class Startup
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="Startup"/> class.
-    /// </summary>
-    /// <param name="configuration">The current configuration.</param>
-    public Startup( IConfiguration configuration ) => Configuration = configuration;
-
-    /// <summary>
-    /// Gets the current configuration.
-    /// </summary>
-    /// <value>The current application configuration.</value>
-    public IConfiguration Configuration { get; }
-
-    /// <summary>
-    /// Configures services for the application.
-    /// </summary>
-    /// <param name="services">The collection of services to configure the application with.</param>
-    public void ConfigureServices( IServiceCollection services )
-    {
-        services.AddControllers();
-        services.AddApiVersioning(
-                    options =>
-                    {
-                        // reporting api versions will return the headers
-                        // "api-supported-versions" and "api-deprecated-versions"
-                        options.ReportApiVersions = true;
-
-                        options.Policies.Sunset( 0.9 )
-                                        .Effective( DateTimeOffset.Now.AddDays( 60 ) )
-                                        .Link( "policy.html" )
-                                            .Title( "Versioning Policy" )
-                                            .Type( "text/html" );
-                    } )
-                .AddMvc()
-                .AddApiExplorer(
-                    options =>
-                    {
-                        // add the versioned api explorer, which also adds IApiVersionDescriptionProvider service
-                        // note: the specified format code will format the version as "'v'major[.minor][-status]"
-                        options.GroupNameFormat = "'v'VVV";
-
-                        // note: this option is only necessary when versioning by url segment. the SubstitutionFormat
-                        // can also be used to control the format of the API version in route templates
-                        options.SubstituteApiVersionInUrl = true;
-                    } );
-        services.AddTransient<IConfigureOptions<SwaggerGenOptions>, ConfigureSwaggerOptions>();
-        services.AddSwaggerGen(
-            options =>
-            {
-                // add a custom operation filter which sets default values
-                options.OperationFilter<SwaggerDefaultValues>();
-
-                // integrate xml comments
-                options.IncludeXmlComments( XmlCommentsFilePath );
-            } );
-    }
-
-    /// <summary>
-    /// Configures the application using the provided builder, hosting environment, and API version description provider.
-    /// </summary>
-    /// <param name="app">The current application builder.</param>
-    /// <param name="provider">The API version descriptor provider used to enumerate defined API versions.</param>
-    public void Configure( IApplicationBuilder app, IApiVersionDescriptionProvider provider )
-    {
-        app.UseRouting();
-        app.UseEndpoints( builder => builder.MapControllers() );
-        app.UseSwagger();
-        app.UseSwaggerUI(
-            options =>
-            {
-                // build a swagger endpoint for each discovered API version
-                foreach ( var description in provider.ApiVersionDescriptions )
-                {
-                    var url = $"/swagger/{description.GroupName}/swagger.json";
-                    var name = description.GroupName.ToUpperInvariant();
-                    options.SwaggerEndpoint( url, name );
-                }
-            } );
-    }
-
-    private static string XmlCommentsFilePath
-    {
-        get
-        {
-            var basePath = PlatformServices.Default.Application.ApplicationBasePath;
-            var fileName = typeof( Startup ).GetTypeInfo().Assembly.GetName().Name + ".xml";
-            return Path.Combine( basePath, fileName );
-        }
-    }
-}
